refactor(gateway): remove backwards-compat shim headers
Twenty shim headers under include/ros2_medkit_gateway/ (managers, plugins/,
providers/, discovery/, http/) were introduced to keep out-of-tree plugin
consumers buildable across the core/ layer split. There are no out-of-tree
consumers - every package that includes these headers lives in our
monorepo - so the indirection layer protected against a scenario that does
not exist. Plugins that legitimately want a public surface use the
core/plugins/, core/providers/, and core/discovery/models/ paths directly.

Migrate every in-monorepo consumer to the canonical paths and delete the
twenty shim files. plugins/plugin_context.hpp consumers are routed to
plugins/ros_plugin_context.hpp (the ROS-aware subclass that defines
node()); ros_plugin_context.hpp is not a shim and stays.
