--- src/ros2_medkit_gateway/include/ros2_medkit_gateway/fault_manager.hpp
@@ -1,20 +1,0 @@
-// Copyright 2026 bburda
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#pragma message("deprecated: include ros2_medkit_gateway/core/managers/fault_manager.hpp instead")
-
-// Backwards-compatibility shim - header relocated to core/managers/.
-#include "ros2_medkit_gateway/core/managers/fault_manager.hpp"
