--- src/ros2_medkit_gateway/include/ros2_medkit_gateway/providers/data_provider.hpp
@@ -1,19 +1,0 @@
-// Copyright 2026 bburda
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-// Backwards-compatibility shim - header moved to core/. Remove once all
-// downstream consumers have migrated to the new path.
-#include "ros2_medkit_gateway/core/providers/data_provider.hpp"
