+++ src/ros2_medkit_plugins/ros2_medkit_graph_provider/src/graph_provider_plugin_exports.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ros2_medkit_gateway/plugins/plugin_types.hpp"
+#include "ros2_medkit_gateway/core/plugins/plugin_types.hpp"
 #include "ros2_medkit_graph_provider/graph_provider_plugin.hpp"
 
 using namespace ros2_medkit_gateway;
