+++ src/ros2_medkit_plugins/ros2_medkit_sovd_service_interface/src/service_exports.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ros2_medkit_gateway/plugins/plugin_types.hpp"
+#include "ros2_medkit_gateway/core/plugins/plugin_types.hpp"
 #include "sovd_service_interface.hpp"
 
 using namespace ros2_medkit_gateway;
